Fix random methods and make usage more consistent
This commit replaces the function `random_i` with `randrange`. This
matches the Python nomenclature of using `randrange` for `[lo, hi)` and
`randint` for `[lo, hi]`, and removes a glaringly obvious possibility
for a divide-by-zero error when `random_i` was called with `lo == hi`.

For the sake of consistency, I also added `randbool` to avoid scattered
calls to `rand()` and usage of `RAND_MAX`.

Note: The current implementation of `randint` is slightly biased when
`hi - lo` is close to `RAND_MAX`, but that really doesn't matter for our
application.