Move to autotools build system
Add `configure.ac` and `Makefile.am` for generating a configure script
and Makefile. To match with convention, source files have been moved
into the `src` subdirectory.

The build files can be generated with `autoreconf -i`. We make use of
some extensions (`AX_*` macros), so we  require the build machone to
have the autoconf archive installed (`autoconf-archive` on Arch Linux).

Note that as many compiler warnings as possible are enabled by default.
For development versions (i.e. commits not tagged with a version string)
of the code, warnings will cause an error. Release versions are slightly
more slack, and will just emit warnings: this is a feature of the
`AX_COMPILER_FLAGS` macro, and is the intended behaviour.