Fixes for old versions of autotools
Bump the required version of autoconf to 2.64. It was previously 2.63,
which is *really* old (2008), and unsupported by the third-party macros
we are using.

This commit also adds the `AM_PROG_CC_C_O` macro, which is now obsolete
but is required by older (e.g. 2.69 as shipped with RHEL7) versions of
autotools.