Strip leading "v" from version tag in configure.ac (#16)
Previously, configure.ac used the latest git tag verbatim as the version
number. Version tags are conventionally prefixed with "v", so this
commit explicitly strips that prefix.

Note the use of "sed" rather than "$SED" in "configure.ac". This is
fine, as maintainer machines can reasonably be expected to have a
sensibly-configured sed.