Implement a colour palette
This commit adds a colour palette. Our "colours" are ncurses "pairs". We
define a palette as an ordered list of pair IDs. We define it like this
because the pair IDs returned by `alloc_pair` are not necessarily
consecutive, and we want to know the relationship between color `i` and
`i+1`.

On terminals that allow it, we redefine the actual colours used by the
terminal. Alternatively, on terminals that support direct colour we use
that.

The default palette is a sweep of HSL space at maximum lightness and
saturation.