Fix width of a color, and care about to signedness
This commit fixes the width of a "color" to a 32-bit unsigned integer.
In practise, this is unlikely to matter: any system this is likely to
compile on is likely to have an int greater than 24 bits. There's no
harm in being more rigorous, though.

At the same time, we also differentiate functions that return a "size"
(e.g. number of colours), and return a size_t rather than an int. Again,
this is unlikely to matter, but there is no harm.