Add error handling routines
Added routines for recording and displaying errors. These are useful
because a simple "fprintf(stderr, ...)" will be obscured by the ncurses
window.

To record an error, call "set_error()" with a "cpipes_errno" as an
argument. A cpipes_errno is just a typedef for an element of the enum
PIPES_C_ERRORS. C doesn't make the distinction, so these can all be
treated as negative integers. Some errors take arguments, which should
be passed into set_error(): see the "PIPES_C_ERROR_STRINGS" array in
"err.c".

Information can be added to an error message to provide contextualising
information as a kind of poor man's stack trace.