Fix call to `attr_set` for ncurses < 6
Ncurses with ABI >= 6 support passing a color index to `attr_set`
through the `opts` pointer. This work-around is because the `params`
pointer must remain a `short` for backwards compatibility and the `opts`
pointer has never actually been used for anything else.

This commit adds an explicit ABI version check that uses the `params`
field when ncurses major version is < 6. I've used an explicit
preprocessor guard  rather than always calling `attr_set(..., p->color,
&p->color)` because `p->color` is an `int`, and it would overflow when
passed as the `short params` argument. Probably not a problem in
practice, but easy to avoid nonetheless.