Add linked list implentation for canvas cells
The plan is to store the list of characters that have been drawn on each
grid cell of the canvas in a linked list. Then, we can erase a character
drawn by a particular pipe by looking backwards from the tail and
popping the first matching pipe cell from the list.

Then, if the list is empty we will erase that cell and if the list still
contains elements, make sure that if the head of the linked list has
changed, it is the character drawn to the terminal.

For example, we might have the following list for a particular cell:

        pipe    char    color
        0x1     ━       1      ← Head
        0x2     ┃       2
        0x3     ━       1
        0x1     ┓       3      ← Tail

If we remove pipe 0x1 from the list, the tail will point to 0x3, and the
head is unchanged so we will not need to change the character drawn at
this cell position. If we then remove 0x1 again, the head will point to
0x2 and a "┃" character must be drawn with color 1. If we remove all
pipes, that cell is erased.