Allow for finite-length pipes
This commit integrates the previous few commits into cpipes.

The function "canvas_erase_tail" is added, which is repsonsible for
removing a pipe from a pipe_cell_list at the tail of a pipe, and then
drawing either a new pipe character, erasing that cell or leaving it
as-is if the uppermost pipe did not change.

The "-m" or "--max" parameter was added to the command line interface to
allow setting the maximum pipe length. This value is now also accepted
by the "init_pipe" function, which now initialises the location_buffer
struct in each pipe.