Implementation of a circular location buffer
The location_buffer struct will be used to store a list of locations for
a pipe in a circular buffer. When a pipe moves forwards, the buffer will
be updated so that the previous tail is replaced by the next location,
and the new head will be pushed.

For a buffer of length 4:

        buffer     pushed
        1               1
        21              2
        321             3
        4321            4
        5432            5
        6543            6

The implementation of this circular buffer is fairly standard. The
buffer itself is a fixed-size portion of memory, and a head index is
updated each time an element is pushed.