Fix segfault/memory errors when resizing terminal
When the --max option is given and terminal is resized, canvas_resize
must be called in order to initialise the pipe cell list in the new
section of the terminal. If the terminal has shrunk, cpipes should not
attempt to pop characters out of cells in the newly-freed memory
region. Those cells are invisible anyway.

For large terminal size changes, wrapping the pipe location by adding or
subtracting the terminal with or height can be insufficient. This commit
changes the behaviour of wrap_pipe to ensure the pipe always ends up in
a valid region of the terminal.