Bugfix: find_num_colors was permuting argv
The "find_num_colors" function counts the number of colors ("-c") given
on the command line. Before this commit, it just used the optstring
"c:", but this would allow getopt to reorder the remaining arguments and
break the next call to getopt using the correct optstring. The sollution
is just to use the same opstring for both calls to getopt.