Less restrictive builds
This is a slight cleanup of the build system, and a couple of trivial
drive-by changes to clean up the source files.

The big change is that `configure`, `config.h.in` and `Makefile.in` are
now included in the source tree. Previously, users who wanted to build
from git needed to have a copy of GNU autotools. Considering I haven't
set up CI to build releases, this is an onerous requirement.

I've also included a Meson build configuration, which is much less
clunky than autotools.